Revert "fix StatelessXmlReporter to generate TestCase.xml problem"
This reverts commit ba16a647

+++ maven-surefire-common/src/main/java/org/apache/maven/plugin/surefire/report/Utf8RecodingDeferredFileOutputStream.java
@@ -93,7 +93,10 @@ final class Utf8RecodingDeferredFileOutputStream
             try
             {
                 close();
-                deferredFileOutputStream.getFile().deleteOnExit();
+                if ( !deferredFileOutputStream.getFile().delete() )
+                {
+                    deferredFileOutputStream.getFile().deleteOnExit();
+                }
             }
             catch ( IOException ioe )
             {
